Fix XO errors, refactor
