+++ lib/dimmer/dimmer.js
@@ -8,7 +8,7 @@ let UUIDGen;
 
 const oldOptKeys = ['dpsOn', 'dpsBright', 'minVal', 'maxVal'];
 function checkDimmerOptionsUpgrade(deviceConfig, log) {
-  const upgOpts = Object.keys(deviceConfig).filter((k) => oldOptKeys.includes(k));
+  const upgOpts = Object.keys(deviceConfig).filter(k => oldOptKeys.includes(k));
   if (upgOpts && upgOpts.length > 0) {
     log.warn(`Please follow the examples in the readme to upgrade these options: ${upgOpts}`);
   }
@@ -32,7 +32,7 @@ class DimmerAccessory {
 
     this.dpsMap = {
       onOff: ('dpsOn' in this.deviceConfig.options) ? this.deviceConfig.options.dpsOn : 1,
-      brightness: ('dpsBright' in this.deviceConfig.options) ? this.deviceConfig.options.dpsBright : 2,
+      brightness: ('dpsBright' in this.deviceConfig.options) ? this.deviceConfig.options.dpsBright : 2
     };
     this.minVal = ('minVal' in this.deviceConfig.options) ? this.deviceConfig.options.minVal : 11;
     this.maxVal = ('maxVal' in this.deviceConfig.options) ? this.deviceConfig.options.maxVal : 244;
@@ -40,13 +40,14 @@ class DimmerAccessory {
     this.device = new TuyaDevice(deviceConfig);
     this.device.connect();
 
-    this.device.on(
-      'data',
-      function(data, command) {
-        if (data !== undefined) {
-          this.log.debug('Updating', this.homebridgeAccessory.displayName, data, command);
-          for (var index in data.dps) {
-            switch (parseInt(index)) {
+    this.device.on('data', (data, command) => {
+      if (data === undefined) {
+        this.log.debug('No data.');
+      } else {
+        this.log.debug('Updating', this.homebridgeAccessory.displayName, data, command);
+        for (const index in data.dps) {
+          if (Object.prototype.hasOwnProperty.call(data.dps, index)) {
+            switch (parseInt(index, 10)) {
               case this.dpsMap.onOff:
                 this.homebridgeAccessory
                   .getService(Service.Lightbulb)
@@ -63,21 +64,18 @@ class DimmerAccessory {
                 this.log.error('Unmapped dps data', index, data.dps[index]);
             }
           }
-        } else {
-          this.log.debug('No date');
         }
-      }.bind(this)
+      }
+    }
     );
 
-    this.device.on(
-      'error',
-      function(data) {
-        this.log.debug('ERROR: not responding', this.homebridgeAccessory.displayName);
-        this.homebridgeAccessory
-          .getService(Service.Lightbulb)
-          .getCharacteristic(Characteristic.On)
-          .updateValue(new Error('Not Responding'));
-      }.bind(this)
+    this.device.on('error', () => {
+      this.log.debug('ERROR: not responding', this.homebridgeAccessory.displayName);
+      this.homebridgeAccessory
+        .getService(Service.Lightbulb)
+        .getCharacteristic(Characteristic.On)
+        .updateValue(new Error('Not Responding'));
+    }
     );
 
     if (this.homebridgeAccessory) {
@@ -163,5 +161,5 @@ class DimmerAccessory {
 
 module.exports = {
   DimmerAccessory,
-  checkDimmerOptionsUpgrade,
+  checkDimmerOptionsUpgrade
 };
