+++ lib/generic/generic.js
@@ -18,16 +18,16 @@ class GenericAccessory {
     this.deviceConfig = deviceConfig;
 
     if (this.deviceConfig.dps) {
-        this.log.warn("Looks like you're using an old config format. Please check the README for an updated example.");
+      this.log.warn('Looks like you\'re using an old config format. Please check the README for an updated example.');
     }
     this.deviceConfig.options = this.deviceConfig.options || {};
 
     if ('dps' in this.deviceConfig.options) {
-    	this.dps = this.deviceConfig.options.dps;
+      this.dps = this.deviceConfig.options.dps;
     } else if ('dps' in this.deviceConfig) {
-    	this.dps = this.deviceConfig.dps;
+      this.dps = this.deviceConfig.dps;
     } else {
-    	this.dps = 1;
+      this.dps = 1;
     }
 
     this.device = new TuyaDevice(deviceConfig);
